CatRentalMachine refactored into separate file
