+++ .replit
@@ -1,5 +1,5 @@
 language = "cpp"
-run = "clang++ -std=c++17 -Wno-deprecated -Wno-padded -Weverything -Wno-c++98-compat -Wno-missing-prototypes -Wno-exit-time-destructors main.cpp Wrappers.cpp Cat.cpp VendingMachine.cpp ItemDispenser.cpp Computer.cpp Drive.cpp CatCyberOverlord.cpp && ./a.out"
+run = "clang++ -std=c++17 -Wno-deprecated -Wno-padded -Weverything -Wno-c++98-compat -Wno-missing-prototypes -Wno-exit-time-destructors main.cpp Wrappers.cpp Cat.cpp VendingMachine.cpp ItemDispenser.cpp Computer.cpp Drive.cpp CatCyberOverlord.cpp CatRentalMachine.cpp && ./a.out"
 hidden = ["ccls"]
 
 [debugger]
